+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * Fix: Integrate up sells cross sells plugin.
     * Fix: Integrate add flat rate plugin.
     * Fix: Improve tax settings.
+    * Fix: Fatal error when ordered product does not exists.
 * 2.0.8 - 2017.02.01:
     * Fix: Allow to remove all attachments.
     * Fix: Do not recalculate discounts for placed orders.